// src/api/openapi/models/EventAction.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type EventAction = {
    action?: EventAction.action;
};

export namespace EventAction {

    export enum action {
        GOING = 'going',
        INTERESTED = 'interested',
    }


}